Gettig list of installed RPM packages
Wnen value of $PACKAGE_SYSTEM is "rpm" then for getting list of installed packages is used command rpm -qa --queryformat="%{NAME}\n". This is good solution because this command returns just list of installed packages and in this case just names of these packages. Each package is evaluated as NOFILES but just for now.